+++ .condarc.yaml
@@ -1,0 +1,5 @@
+channels:
+  - defaults
+  - bioconda
+  - conda-forge
+show_channels_urls: True
